+++ src/main/java/frc/robot/Constants.java
@@ -119,6 +119,23 @@ public final class Constants {
         public static final double kITurning = 0.;
     }
 
+    public final class Buttons {
+        public static final int A = 1;
+        public static final int B = 2;
+        public static final int X = 3;
+        public static final int Y = 4;
+        public static final int LEFT_BUMPER = 5;
+        public static final int RIGHT_BUMPER = 6;
+        public static final int VIEW = 7;
+        public static final int MENU = 8;
+        public static final int LEFT_STICK_BUTTON = 9;
+        public static final int RIGHT_STICK_BUTTON = 10;
+        public static final int POV_UP = 0;
+        public static final int POV_DOWN = 180;
+        public static final int POV_RIGHT = 90;
+        public static final int POV_LEFT = 270;
+    }
+
     public static final class DriveConstants {
         // left-to-right distance between the drivetrain wheels, should be measured from center to center AND IN METERS
         public static final double kTrackWidth = 0.5715;
