+++ vendordeps/ThriftyLib.json
@@ -1,0 +1,20 @@
+{
+    "fileName": "ThriftyLib.json",
+    "name": "ThriftyLib",
+    "version": "2025.1.1",
+    "frcYear": "2025",
+    "uuid": "60b2694b-9e6e-4026-81ee-6f167946f4b0",
+    "mavenUrls": [
+        "https://docs.home.thethriftybot.com"
+    ],
+    "jsonUrl": "https://docs.home.thethriftybot.com/ThriftyLib.json",
+    "javaDependencies": [
+        {
+            "groupId": "com.thethriftybot.frc",
+            "artifactId": "ThriftyLib-java",
+            "version": "2025.1.1"
+        }
+    ],
+    "jniDependencies": [],
+    "cppDependencies": []
+}
