+++ vendordeps/maple-sim.json
@@ -1,0 +1,26 @@
+{
+    "fileName": "maple-sim.json",
+    "name": "maplesim",
+    "version": "0.3.11prev",
+    "frcYear": "2025",
+    "uuid": "c39481e8-4a63-4a4c-9df6-48d91e4da37b",
+    "mavenUrls": [
+        "https://shenzhen-robotics-alliance.github.io/maple-sim/vendordep/repos/releases",
+        "https://repo1.maven.org/maven2"
+    ],
+    "jsonUrl": "https://shenzhen-robotics-alliance.github.io/maple-sim/vendordep/maple-sim.json",
+    "javaDependencies": [
+        {
+            "groupId": "org.ironmaple",
+            "artifactId": "maplesim-java",
+            "version": "0.3.11prev"
+        },
+        {
+            "groupId": "org.dyn4j",
+            "artifactId": "dyn4j",
+            "version": "5.0.2"
+        }
+    ],
+    "jniDependencies": [],
+    "cppDependencies": []
+}
